Make POST requests use external-format-out for multipart/form-data.
With this change, POST requests that use multipart/form-data will use
external-format-out for encoding, instead of :latin-1. Without this
change, an attempt to use characters outside the :latin-1 range (e.g,
U+2011, "n-dash") would cause an exception while building the request.

Note: it would be possible to only do this for strings that have
characters that are not standard-char.