Refactor time zone parsing into regular expression
This is in response to pull request #62 which proposed special casing
for an all-numeric offset in an incorrect way.  Offsets may now be
specified with or without a colon between the hour and the minute.
The "GMT" prefix is now optional to cater for some broken servers.
All HTTP date specifications need to be in GMT according to the spec,
so all the time zone parsing business is catering for broken servers
anyway.